Add Core Length Score, and UART Fix
Added feature to algorithms to score average length
of the core_anchors in hopes of higher accuracy.

Fixed Serial Communication as well, so the
different cameras send proper coordinates to the
microcontroller.

As of right now the laser detection for green laser is
pretty solid. Need to work on red laser.