Finalize Project with README updates
Used previous commit at expo, with some minor modifications
to tuning (still on turret unfortunately, but not saved in
this commit).

Went to Senior Project Demo / Expo, got first place, and
absolutely killed it. Super proud of this project, and I
hope future ones can be as good / surpass this one :)